chore: convert zoeken.model package to Kotlin
