+++ src/main/kotlin/net/atos/zac/zoeken/model/DocumentIndicatie.kt
@@ -1,0 +1,16 @@
+package net.atos.zac.zoeken.model
+
+/*
+ * This enum represents the possible indicaties for a informatieobject.
+ *
+ * The order of the enum values determines the sort priority of the indicators (the highest priority indicator first)
+ * Nota bene: When the order of the enum values changes, the Solr index for zaken MUST be rebuilt.
+ * Oh, and no more than 63 indicaties in this enum please (it needs to fit in a signed plong in Solr)
+ */
+enum class DocumentIndicatie {
+    VERGRENDELD,
+    ONDERTEKEND,
+    BESLUIT,
+    GEBRUIKSRECHT,
+    VERZONDEN
+}
